[repo] Flatten ast folder (part 3)
