+++ samlang-core/src/ast/__tests__/common-nodes.test.ts
@@ -14,6 +14,9 @@ import {
   tupleType,
   functionType,
   UndecidedTypes,
+  Position,
+  Range,
+  ModuleReference,
 } from '../common-nodes';
 
 it('Literals have expected pretty printed values', () => {
@@ -118,4 +121,122 @@ it('type equality test', () => {
   expect(
     isTheSameType({ type: 'UndecidedType', index: 0 }, { type: 'UndecidedType', index: 0 })
   ).toBeTruthy();
+});
+
+it('Correct toString() for display', () => {
+  expect(Position.DUMMY.toString()).toBe('0:0');
+  expect(new Position(42, 53).toString()).toBe('43:54');
+});
+
+it('Position comparison smoke test', () => {
+  expect(new Position(1, 4).compareTo(new Position(1, 4))).toBe(0);
+});
+
+it('Correct same line comparision test', () => {
+  expect(new Position(1, 2).compareTo(new Position(1, 3))).toBeLessThan(0);
+  expect(new Position(1, 3).compareTo(new Position(1, 2))).toBeGreaterThan(0);
+  expect(new Position(1, 2).compareTo(new Position(1, 3))).toBeLessThan(0);
+  expect(new Position(1, 3).compareTo(new Position(1, 2))).toBeGreaterThan(0);
+  expect(new Position(1, 4).compareTo(new Position(1, 3))).toBeGreaterThan(0);
+  expect(new Position(1, 3).compareTo(new Position(1, 4))).toBeLessThan(0);
+  expect(new Position(1, 4).compareTo(new Position(1, 3))).toBeGreaterThan(0);
+  expect(new Position(1, 3).compareTo(new Position(1, 4))).toBeLessThan(0);
+});
+
+it('Correct different lines comparision test', () => {
+  expect(new Position(1, 2).compareTo(new Position(2, 3))).toBeLessThan(0);
+  expect(new Position(1, 3).compareTo(new Position(2, 3))).toBeLessThan(0);
+  expect(new Position(1, 4).compareTo(new Position(2, 3))).toBeLessThan(0);
+  expect(new Position(2, 3).compareTo(new Position(1, 2))).toBeGreaterThan(0);
+  expect(new Position(2, 3).compareTo(new Position(1, 3))).toBeGreaterThan(0);
+  expect(new Position(2, 3).compareTo(new Position(1, 4))).toBeGreaterThan(0);
+});
+
+it('Range.toString() works as expected', () => {
+  expect(Range.DUMMY.toString()).toBe('0:0-0:0');
+  expect(new Range(new Position(1, 1), new Position(2, 4)).toString()).toBe('2:2-3:5');
+});
+
+it('Range.containsPosition() works as expected', () => {
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsPosition(new Position(2, 2))
+  ).toBeTruthy();
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsPosition(new Position(1, 2))
+  ).toBeFalsy();
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsPosition(new Position(3, 2))
+  ).toBeFalsy();
+});
+
+it('Range.containsRange() works as expected', () => {
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsRange(
+      new Range(new Position(1, 3), new Position(3, 1))
+    )
+  ).toBeTruthy();
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsRange(
+      new Range(new Position(1, 4), new Position(3, 0))
+    )
+  ).toBeTruthy();
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsRange(
+      new Range(new Position(1, 3), new Position(3, 2))
+    )
+  ).toBeFalsy();
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsRange(
+      new Range(new Position(1, 2), new Position(3, 1))
+    )
+  ).toBeFalsy();
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1)).containsRange(
+      new Range(new Position(1, 2), new Position(3, 2))
+    )
+  ).toBeFalsy();
+});
+
+it('Range.union() works as expected', () => {
+  expect(
+    new Range(new Position(1, 3), new Position(3, 1))
+      .union(new Range(new Position(2, 3), new Position(4, 1)))
+      .toString()
+  ).toBe(new Range(new Position(1, 3), new Position(4, 1)).toString());
+
+  expect(
+    new Range(new Position(2, 3), new Position(4, 1))
+      .union(new Range(new Position(1, 3), new Position(3, 1)))
+      .toString()
+  ).toBe(new Range(new Position(1, 3), new Position(4, 1)).toString());
+
+  expect(
+    new Range(new Position(1, 3), new Position(2, 3))
+      .union(new Range(new Position(3, 1), new Position(4, 1)))
+      .toString()
+  ).toBe(new Range(new Position(1, 3), new Position(4, 1)).toString());
+
+  expect(
+    new Range(new Position(3, 1), new Position(4, 1))
+      .union(new Range(new Position(1, 3), new Position(2, 3)))
+      .toString()
+  ).toBe(new Range(new Position(1, 3), new Position(4, 1)).toString());
+});
+
+it('ModuleReference.toString()', () => {
+  expect(ModuleReference.ROOT.toString()).toBe('');
+  expect(new ModuleReference(['Foo']).toString()).toBe('Foo');
+  expect(new ModuleReference(['Foo', 'Bar']).toString()).toBe('Foo.Bar');
+});
+
+it('ModuleReference.toFilename', () => {
+  expect(ModuleReference.ROOT.toFilename()).toBe('.sam');
+  expect(new ModuleReference(['Foo']).toFilename()).toBe('Foo.sam');
+  expect(new ModuleReference(['Foo', 'Bar']).toFilename()).toBe('Foo/Bar.sam');
+});
+
+it('ModuleReference.uniqueHash is ModuleReference.toString', () => {
+  expect(new ModuleReference(['Foo', 'Bar']).toString()).toBe(
+    new ModuleReference(['Foo', 'Bar']).uniqueHash()
+  );
 });
