+++ samlang-core/src/checker/__tests/global-typing-context-builder.test.ts
@@ -1,6 +1,4 @@
-import { functionType, intType } from '../../ast/common-nodes';
-import ModuleReference from '../../ast/common/module-reference';
-import Range from '../../ast/common/range';
+import { functionType, intType, Range, ModuleReference } from '../../ast/common-nodes';
 import { EXPRESSION_FALSE } from '../../ast/samlang-expressions';
 import type { ClassDefinition, SamlangModule } from '../../ast/samlang-toplevel';
 import { mapOf } from '../../util/collections';
