+++ samlang-core/src/checker/__tests/index.test.ts
@@ -4,7 +4,7 @@ import {
   // eslint-disable-next-line camelcase
   typeCheckSingleModuleSource_EXPOSED_FOR_TESTING,
 } from '..';
-import ModuleReference from '../../ast/common/module-reference';
+import { ModuleReference } from '../../ast/common-nodes';
 import { createGlobalErrorCollector } from '../../errors';
 import { parseSamlangModuleFromText } from '../../parser';
 import { mapOf, hashMapOf } from '../../util/collections';
