+++ samlang-core/src/checker/__tests/undefined-import-checker.test.ts
@@ -1,5 +1,4 @@
-import ModuleReference from '../../ast/common/module-reference';
-import Range from '../../ast/common/range';
+import { ModuleReference, Range } from '../../ast/common-nodes';
 import type { Sources } from '../../ast/common/structs';
 import { SamlangModule, ClassDefinition } from '../../ast/samlang-toplevel';
 import { createGlobalErrorCollector } from '../../errors';
