+++ samlang-core/src/checker/constraint-aware-checker.ts
@@ -1,5 +1,4 @@
-import { Type, UndecidedType, isTheSameType } from '../ast/common-nodes';
-import type Range from '../ast/common/range';
+import { Type, UndecidedType, isTheSameType, Range } from '../ast/common-nodes';
 import { ModuleErrorCollector } from '../errors';
 import TypeResolution from './type-resolution';
