+++ samlang-core/src/compiler/hir/__tests__/index.test.ts
@@ -5,10 +5,10 @@ import {
   identifierType,
   functionType,
   unitType,
+  Range,
+  ModuleReference,
 } from '../../../ast/common-nodes';
 import { MUL, MINUS, EQ } from '../../../ast/common-operators';
-import ModuleReference from '../../../ast/common/module-reference';
-import Range from '../../../ast/common/range';
 import {
   HIR_VARIABLE,
   HIR_RETURN,
