+++ samlang-core/src/compiler/mir/index.ts
@@ -1,5 +1,5 @@
 import { ENCODED_COMPILED_PROGRAM_MAIN, encodeMainFunctionName } from '../../ast/common-names';
-import type ModuleReference from '../../ast/common/module-reference';
+import type { ModuleReference } from '../../ast/common-nodes';
 import type { Sources, GlobalVariable } from '../../ast/common/structs';
 import type { HighIRModule } from '../../ast/hir-toplevel';
 import {
