+++ samlang-core/src/errors/__tests__/index.test.ts
@@ -15,9 +15,7 @@ import {
   DuplicateFieldDeclarationError,
   NonExhausiveMatchError,
 } from '..';
-import { intType, boolType } from '../../ast/common-nodes';
-import ModuleReference from '../../ast/common/module-reference';
-import Range from '../../ast/common/range';
+import { intType, boolType, Range, ModuleReference } from '../../ast/common-nodes';
 
 const testCases: readonly (readonly [CompileTimeError, string])[] = [
   [
