+++ samlang-core/src/errors/index.ts
@@ -1,8 +1,6 @@
 /* eslint-disable @typescript-eslint/no-empty-interface */
 
-import { Type, prettyPrintType } from '../ast/common-nodes';
-import type ModuleReference from '../ast/common/module-reference';
-import type Range from '../ast/common/range';
+import { Type, prettyPrintType, Range, ModuleReference } from '../ast/common-nodes';
 
 export abstract class CompileTimeError<T = string> {
   constructor(
