+++ samlang-core/src/parser/__tests__/index.test.ts
@@ -1,5 +1,5 @@
 import { parseSamlangModuleFromText, parseSamlangExpressionFromText } from '..';
-import ModuleReference from '../../ast/common/module-reference';
+import { ModuleReference } from '../../ast/common-nodes';
 import type { SamlangExpression } from '../../ast/samlang-expressions';
 import { createGlobalErrorCollector } from '../../errors';
