+++ samlang-core/src/parser/parser-class-builder.ts
@@ -1,7 +1,6 @@
 import { AbstractParseTreeVisitor } from 'antlr4ts/tree/AbstractParseTreeVisitor';
 
-import { functionType } from '../ast/common-nodes';
-import Range from '../ast/common/range';
+import { functionType, Range } from '../ast/common-nodes';
 import type { TypeDefinition } from '../ast/common/structs';
 import { SamlangExpression } from '../ast/samlang-expressions';
 import { ClassDefinition, ClassMemberDefinition } from '../ast/samlang-toplevel';
